Updated with Share and Rate App Functionality
