+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 23
     buildToolsVersion '23.0.3'
     defaultConfig {
-        applicationId 'com.androidbuts.saveinsta'
+        applicationId 'com.androidbuts.instasaver'
         minSdkVersion 15
         targetSdkVersion 23
         versionCode 1
