+++ app/src/main/java/com/androidbuts/saveinsta/model/EventChangeDominantColor.java
@@ -1,0 +1,7 @@
+package com.androidbuts.saveinsta.model;
+
+/**
+ * Created by Pratik Butani
+ */
+public class EventChangeDominantColor {
+}
