+++ app/src/main/java/com/androidbuts/saveinsta/model/EventVideoIsDownload.java
@@ -1,0 +1,17 @@
+package com.androidbuts.saveinsta.model;
+
+/**
+ * Created by Created by Pratik Butani
+ */
+public class EventVideoIsDownload {
+
+    private boolean isDownload;
+
+    public EventVideoIsDownload(boolean isDownload) {
+        this.isDownload = isDownload;
+    }
+
+    public boolean isDownload() {
+        return isDownload;
+    }
+}
