+++ app/src/main/java/com/androidbuts/saveinsta/model/InstaData.java
@@ -1,0 +1,13 @@
+package com.androidbuts.saveinsta.model;
+
+/**
+ * Created by Pratik Butani
+ */
+public class InstaData {
+
+    private InstaEntryData entry_data;
+
+    public InstaEntryData getEntryData() {
+        return entry_data;
+    }
+}
