--- app/src/main/java/com/mikhaellopez/saveinsta/model/EventChangeDominantColor.java
@@ -1,7 +1,0 @@
-package com.mikhaellopez.saveinsta.model;
-
-/**
- * Created by Pratik Butani
- */
-public class EventChangeDominantColor {
-}
