--- app/src/main/java/com/mikhaellopez/saveinsta/model/EventVideoIsDownload.java
@@ -1,17 +1,0 @@
-package com.mikhaellopez.saveinsta.model;
-
-/**
- * Created by Created by Pratik Butani
- */
-public class EventVideoIsDownload {
-
-    private boolean isDownload;
-
-    public EventVideoIsDownload(boolean isDownload) {
-        this.isDownload = isDownload;
-    }
-
-    public boolean isDownload() {
-        return isDownload;
-    }
-}
