--- app/src/main/java/com/mikhaellopez/saveinsta/model/InstaData.java
@@ -1,13 +1,0 @@
-package com.mikhaellopez.saveinsta.model;
-
-/**
- * Created by Pratik Butani
- */
-public class InstaData {
-
-    private InstaEntryData entry_data;
-
-    public InstaEntryData getEntryData() {
-        return entry_data;
-    }
-}
