+++ app/src/main/res/menu/menu_main.xml
@@ -4,11 +4,16 @@
     tools:context=".controller.activity.MainActivity">
 
     <item
+        android:id="@+id/action_share"
+        android:icon="@drawable/ic_share_white_24dp"
+        android:title="@string/string_share_app"
+        app:showAsAction="ifRoom" />
+
+    <item
         android:id="@+id/action_info"
-        android:icon="@drawable/ic_info"
-        android:menuCategory="secondary"
-        android:orderInCategory="1"
+        android:icon="@drawable/ic_help_outline_white_24dp"
         android:title="@string/string_information"
         app:showAsAction="ifRoom" />
+
 
 </menu>
