--- app/src/main/res/values-fr/strings.xml
@@ -1,25 +1,0 @@
-<resources>
-    <string name="app_name">SaveInsta</string>
-    <string name="edit_text_hint">Colle URL de photo/vidéo Instagram ici</string>
-
-    <string name="info_message">
-        3 Étapes pour télécharger une photo/vidéo Instagram\n\n
-        1. Ouvrir l\'application Instagram, choisir Copier l\'URL de publication sur la photo/vidéo que vous voulez sauvegarder\n\n
-        2. Ouvrir SaveInstagram et laisser la magie faire le reste\n\n
-        3. Télécharger la Photo/Vidéo en cliquant sur le bouton sauvegarder en bas à droite de SaveInstagram ou cliquer sur la croix en haut à droite pour la supprimer. Ouvrir la galerie pour voir le résultat!
-    </string>
-
-    <string name="dialog_button_rate_app">NOTE CETTE APP</string>
-    <string name="snackbar_no_market">Vous n\'avez pas de Market App</string>
-    <string name="dialog_message_why_permission">Vous devez activer cette permission pour télécharger la photo</string>
-    <string name="snackbar_no_instagram">Vous n\'avez pas Instagram</string>
-    <string name="snackbar_no_insta_url">Ce n\'est pas une URL de photo/vidéo Instagram</string>
-    <string name="snackbar_input_empty">Input vide</string>
-    <string name="snackbar_load_photo_failed">Chargement photo échoué</string>
-    <string name="snackbar_photo_saved">Photo sauvé!</string>
-    <string name="snackbar_save_photo_failed">Sauvegarde photo échoué</string>
-    <string name="snackbar_video_saved">Vidéo sauvé!</string>
-    <string name="snackbar_save_video_failed">Sauvegarde vidéo échoué</string>
-    <string name="button_open_instagram">Ouvrir Instagram</string>
-    <string name="button_load_it">Charger</string>
-</resources>
